cfl.py: Explicitly change array ordering for numpy
Up to now, cfl.py specified Fortran-order (order='F') for the numpy
arrays it created from cfl files. This mostly worked as expected, but
lead to some strange issues, as numpy prefers C-order.

With this commit, the array is rearranged to C-order on reading and
rearranged to Fortran-order on writing. This makes reading and writing
cfl-files slower, but ensures predictable behaviour in numpy.