num/ops.c: replace assert(0) with proper error()
gpuwrp_fun had an assert(0) if not both CUDA and OpenMP were found.
Using CMake with CUDA, which by default does not use OpenMP but does
define NDEBUG this would lead to unwanted behavior: Since NDEBUG was
defined, assert(0) would not trigger an error. Still, the function would
do nothing useful, leading to all-zero results.

This commit leads to an error message and an abort in that case.