fix memcache_clear
Since 25d789ca8a2f7281f9f33a718fec7dc9a0882241, memcache_clear could not
work: it calls free_check_p with a size of 0 to find all free pointers
on a device. With the check introduced in the commit above (rptr->len <=
4 * size), this would not find any pointer at all. Therefore,
memcache_clear did not in fact free any pointers.

Therefore, we now only perform that check if size != 0.

Additionally, we cannot check if the ptr is on the GPU in
cuda_free_wrapper(), as it is already removed from the memcache list at
that time.

The check is unneeded anyway, as the normal cuda_free already performs
that check and the only other place from which it is called is
memcache_clear, which also makes sure to only remove GPU pointers.